+++ packages/atomic/src/components/commerce/product-template-components/atomic-product-children/select-child-product-event.ts
@@ -1,0 +1,5 @@
+import {ChildProduct} from '@coveo/headless/commerce';
+
+export interface SelectChildProductEventArgs {
+  child: ChildProduct;
+}